create and modif blog ok

+++ app/api/protected/delete-post/route.ts
@@ -1,9 +1,12 @@
-import prisma from '../../../lib/prisma';
+import apiAdminAuth from '@/lib/apiAdminAuth';
+import prisma from '../../../../lib/prisma';
 import { NextRequest } from 'next/server';
 
 
 
 async function handler(req:NextRequest) {
+    const auth = await apiAdminAuth()
+    if (!auth)  return Response.json("Unauthorized",{status : 401})
     const reqData = await req.json()
     const { id } = reqData
     const result = await prisma.post.delete({
